Small bug fix

+++ Trace/Shape.cs
@@ -122,7 +122,7 @@ namespace Trace
         private static Vec2D _spherePointToUV(Point point)
             => new Vec2D(
                     (float)Math.Atan2(point.y, point.x) / (2.0f * Constant.PI),
-                    Math.Acos(point.z) / Constant.PI
+                    (float)Math.Acos(point.z) / Constant.PI
                 );
     }
 }
